[Upd #286] Refactor for readability
